backend/api/mixins.py
from rest_framework import mixins, viewsets


class ListCreateRetriveViewSet(
    mixins.CreateModelMixin,
    mixins.ListModelMixin,
    mixins.RetrieveModelMixin,
    viewsets.GenericViewSet
):
    pass
